convert B&W version to two unstroked paths
Previously, the Black and White version of Tux contained
many elements, including clipped paths, Inkscape spiro
paths, and stroked paths. This made it difficult if you
wanted to use tux to print, or needed the white areas to
be the background colour.

This PR converts tux to two paths, the black path of Tux,
and the underneath white part, that can be removed if you
want the cutout effect.